Add new API GSettingsBackendChangeset
This is an import of the DConfChangeset API.  It will eventually be used
as the primary way of communicating changesets between GSettings and
various GSettingsBackend implementations.

It will also be useful in simplifying the implementation of several
GSettingsBackend implementations contained within GLib as well as some
new ones being authored outside of it.

GSettingsBackend has always more or less represented a 'dconf-like'
database structure, which is why DConfChangeset is a generally useful
thing to have as part of the GSettingsBackend API.